Rename logging directory (#3860)
* Updating name of logging package to avoid overwriting Python builtin logging.

* Updating name of logging package to avoid overwriting Python builtin logging.
