+++ official/utils/logs/hooks_helper_test.py
@@ -23,7 +23,7 @@ import unittest
 
 import tensorflow as tf  # pylint: disable=g-bad-import-order
 
-from official.utils.logging import hooks_helper
+from official.utils.logs import hooks_helper
 
 
 class BaseTest(unittest.TestCase):
