+++ official/utils/logs/hooks_test.py
@@ -24,7 +24,7 @@ import time
 import tensorflow as tf  # pylint: disable=g-bad-import-order
 from tensorflow.python.training import monitored_session  # pylint: disable=g-bad-import-order
 
-from official.utils.logging import hooks
+from official.utils.logs import hooks
 
 
 tf.logging.set_verbosity(tf.logging.ERROR)
