+++ official/utils/logs/logger_test.py
@@ -26,7 +26,7 @@ import unittest
 
 import tensorflow as tf  # pylint: disable=g-bad-import-order
 
-from official.utils.logging import logger
+from official.utils.logs import logger
 
 
 class BenchmarkLoggerTest(tf.test.TestCase):
