+++ official/utils/logs/metric_hook.py
@@ -18,9 +18,9 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import tensorflow as tf
+import tensorflow as tf  # pylint: disable=g-bad-import-order
 
-from official.utils.logging import logger
+from official.utils.logs import logger
 
 
 class LoggingMetricHook(tf.train.LoggingTensorHook):
